+++ app/controllers/committees_controller.rb
@@ -5,17 +5,21 @@ class CommitteesController < ApplicationController
   before_action :authenticate_viewer_or_current_user!
   before_action :set_committee, only: :show
 
-  def show
-  end
+  layout "layouts/full_page_sidebar"
 
+  # # GET /committees/1
+  # def show
+  # end
+
+  # GET /committees
   def index
-    @order = scrub_order(Member, params[:order], 'members.last_name')
+    @order = scrub_order(Member, params[:order], "members.last_name")
   end
 
   private
 
   def set_committee
-    @committee = Committee.current.find_by_slug params[:committee]
+    @committee = Committee.current.find_by(slug: params[:committee])
     empty_response_or_root_path(committees_path) unless @committee
   end
 end