+++ app/controllers/reports_controller.rb
@@ -5,6 +5,8 @@ class ReportsController < ApplicationController
   before_action :authenticate_viewer_or_current_user!
   before_action :load_recruitment
 
+  layout "layouts/full_page_sidebar"
+
   def data_quality
     @data_quality = @recruitment.dig(:data_quality) if @recruitment
   end